requirements.txt
alembic==0.9.9
asn1crypto==0.24.0
astroid==1.6.2
autopep8==1.3.5
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.5.1
coveralls==1.3.0
cryptography==1.9
docopt==0.6.2
Flask==0.12.2
Flask-API==1.0
Flask-Cors==3.0.3
Flask-JWT==0.3.2
Flask-JWT-Extended==3.7.2
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gunicorn==19.7.1
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mysqlclient==1.3.12
nose==1.3.7
pycodestyle==2.3.1
pycparser==2.18
PyJWT==1.4.2
pylint==1.8.3
python-dateutil==2.7.2
python-editor==1.0.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.6
typing==3.5.3.0
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
